[FEATURE] Add capability to rewrite links processed via TypoLinkCodecService
Not all links are created via the PageLinkHandler, thus we need an option to rewrite links created through viewhelpers and when parsing RTE content.
The TypoLinkCodecService seems to be the central hub for generation of such links.
This feature adds the same link rewriting capabilities used in the PageLinkHandler override to an XCLASS of TypoLinkCodecServce.
The feature can be turned off in the extension configuration.